Finish LookAt, projection matrix, WIP when to add back face culling, FPS camera

+++ Src/Include/QApp.h
@@ -23,7 +23,7 @@ public:
     void Start();
     void Shutdown();
 
-    void ShowFrameStatistics(int frameCnt, double dt);
+    void ShowFrameStatistics(int frameCnt, float dt);
 
 private:
     bool m_isPaused;
